Did some further refactoring
