+++ JANGAN/ExperimentQueue/BaseRunConfig.ini
@@ -64,7 +64,7 @@ LatestCheckpointPath = "../../Data/BaseExperiment/checkpoints/baseRunConfigRef.t
 # Save the model of the CGAN pr epoch
 SaveCheckpoints = True
 # If a checkpoint exist, use it, if not just train
-UseSavedModel = False
+UseSavedModel = True
 # Directory of log output
 LogPath = "../../Data/BaseExperiment/{TIMESTAMP}/Logs/"
 # How much of the dataset should be training and testing data
@@ -121,5 +121,7 @@ LRScheduler = "Constant"
 LearningRateClassifier = 0.0003
 
 [ClassifierOutput]
+# Path of the output logs
+OutputDir = "../../Data/BaseExperiment/{TIMESTAMP}/Logs/"
 # Path of the classification images
 ClassifyDir = "../../Data/BaseExperiment/{TIMESTAMP}/FakeOutput/"
