+++ JANGAN/JANGAN.pyproj
@@ -70,6 +70,7 @@
     </Compile>
     <Compile Include="ProjectTools\AutoPackageInstaller.py" />
     <Compile Include="ProjectTools\BaseKerasModelTrainer.py" />
+    <Compile Include="ProjectTools\BaseMLModel.py" />
     <Compile Include="ProjectTools\ConfigHelper.py" />
     <Compile Include="ProjectTools\HelperFunctions.py">
       <SubType>Code</SubType>
